Split up code into multiple modules
