+++ bokehjs/src/lib/models/glyphs/webgl/markers.ts
@@ -1,4 +1,4 @@
-import {Program, VertexBuffer, IndexBuffer} from "./gloo"
+import {Program, VertexBuffer, IndexBuffer} from "./utils"
 import {BaseGLGlyph, Transform, attach_float, attach_color} from "./base"
 import {vertex_shader} from "./markers.vert"
 import {fragment_shader} from "./markers.frag"
